rel to #11520: make variable list usable in waypoint calc dialogs
