+++ main/res/layout/cachedetail_variables_page.xml
@@ -112,29 +112,10 @@
             android:layout_marginTop="3dp" />
 
 
-        <!-- Wrapping RecyclerView with a RelativeLayout is necessary for "wrap_content" to
-       work properly. See e.g. https://stackoverflow.com/questions/27475178/how-do-i-make-wrap-content-work-on-a-recyclerview -->
-        <RelativeLayout
+        <cgeo.geocaching.ui.VariableListView
+            android:id="@+id/variables"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-
-            <androidx.recyclerview.widget.RecyclerView
-
-                android:id="@+id/variables_list"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:clipToPadding="false"
-                android:footerDividersEnabled="false"
-                android:headerDividersEnabled="false"
-                android:nestedScrollingEnabled="false"
-                android:orientation="vertical"
-                android:paddingHorizontal="4dip"
-                android:scrollbars="none"
-                tools:ignore="NestedScrolling"
-                tools:listitem="@layout/variables_list_item">
-
-            </androidx.recyclerview.widget.RecyclerView>
-        </RelativeLayout>
+            android:layout_height="wrap_content"/>
 
     </LinearLayout>
