+++ main/res/layout/variable_list_item.xml
